Suggestion: Document notebook as `sub-article`
Proposal: 

Perhaps we should propose that the even a notebook that represents the source of the article itself be included as a `sub-article` (and `xref`ed as appropriate). 

One of the great things about the sub-article approach is that it makes the source code that produces elements of the article available to reviewers or the manuscript system. In the case of a single notebook that is used to produce the article itself, the rendered JATS of the notebook may not include source code (and will have the narrative type structure). If we additionally include a `notebook` style rendering of the source notebook as a `sub-article`, the source code and computional version of the article could be viewed / utilized as well.

This approach would allow us to focus on producing

- The JATS article itself to be a narrative focused rendering of JATS with `xrefs` to `sub-article`s. It will likely do things like exclude the code that generates figures in order to provide a more typical / usable article rendering.
- The `sub-article`s that represent the notebooks (as described in the next document), and including `self-uri`, `supplementary-materials` and so on that provide all the notebook metadata (including ultimately the execution information).

What do you think?